Create githook to ensure correct filename formatting
Git hooks are normally stored at `{project_dir}/.git/hooks/`, which is
not shared. To circumvent this, a `.githooks` folder is created in the
root of the project folder.

Because of this approach, the hooks are not ran automatically. Note that
this is a good thing: you would not want arbitrary hook-code to be
executed without your explicit permission.

To enable the hooks if you are using git version >= 2.9:

	```git config core.hooksPath .hooks```

Otherwise, you could create a symlink or copy the hooks.